chore(ci) do not reuse the Travis cache across branches
This should avoid having `next` or PRs based upon it to use dependencies
built for `master`.

Hopefully this will address the issues we're currently having with the
`pdk` testsuite failing due to `set_real_ip` patch for the stream
module being present in the `next` build of OpenResty and absent
in `master`.